+++ src/Sentry/Platforms/Native/Sentry.Native.props
@@ -1,19 +1,4 @@
 <Project>
 
-<!-- TODO: Only add on AOT builds?-->
-  <ItemGroup>
-
-    <SentryNativeSdk Include="$(MSBuildThisFileDirectory)..\..\..\..\lib\sentry-native\macos\bridge.dylib" />
-    <SentryNativeSdk Include="$(MSBuildThisFileDirectory)..\..\..\..\lib\sentry-native\macos\Sentry\Sentry.dylib" />
-    <SentryNativeSdk Include="$(MSBuildThisFileDirectory)..\..\..\..\lib\sentry-native\macos\Sentry\Sentry.dylib.dSYM" />
-
-    <None Include="@(SentryNativeSdk)">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <Pack>true</Pack>
-      <Link>Platforms/Native/macOS/Sentry.dylib</Link>
-      <PackagePath>native\macOS</PackagePath>
-    </None>
-
-  </ItemGroup>
 
 </Project>
