+++ src/Sentry/Platforms/macOS/SentryOptions.cs
@@ -1,0 +1,9 @@
+namespace Sentry;
+
+public partial class SentryOptions
+{
+    /// <summary>
+    /// Enables native macOS crash reporting.
+    /// </summary>
+    public bool EnableNativeCrashReporting { get; set; }
+}
